Move iOS observability support into package
